fix(windows): clippy lints

+++ crate/pkcs11/proteccio/src/proteccio.rs
@@ -226,7 +226,7 @@ impl From<CK_INFO> for Info {
                 .unwrap_or_default()
                 .to_string_lossy()
                 .to_string(),
-            flags: info.flags,
+            flags: u64::from(info.flags),
             libraryDescription: CStr::from_bytes_until_nul(&info.libraryDescription)
                 .unwrap_or_default()
                 .to_string_lossy()
